refactor: rename closeWatcher to closeHandler
