+++ src/content/constants.ts
@@ -1,0 +1,2 @@
+export const CLOSE_EXISTING_CHECK_DELAY_MS = 100;
+export const CLOSE_DETECTION_TIMEOUT_MS = 10000;
